+++ BambooTracker/gui/order_list_editor/order_list_editor.hpp
@@ -42,6 +42,8 @@ public:
 	int getRowsFontSize() const;
 	void setFonts(QString headerFont, int headerSize, QString rowsFont, int rowsSize);
 
+	void setHorizontalScrollMode(bool cellBased, bool refresh = true);
+
 signals:
 	void currentTrackChanged(int num);
 	void currentOrderChanged(int num);
@@ -75,12 +77,16 @@ public slots:
 
 protected:
 	bool eventFilter(QObject *watched, QEvent *event) override;
+	void showEvent(QShowEvent* event) override;
 
 private:
 	Ui::OrderListEditor *ui;
 	std::shared_ptr<BambooTracker> bt_;
 
 	bool freezed_;
+
+	bool hScrollCellMove_;
+	void updateHorizontalSliderMaximum();
 };
 
 #endif // ORDER_LIST_EDITOR_HPP